+++ src/readme.spec.js
@@ -3,10 +3,17 @@ import {isEqualObjects} from '../spec/helpers/helpers.js'
 
 describe('readme', () => {
 	let debug;
+	let facts;
+	let dimensionHierarchies;
+	let cube;
+	let regions;
+	let date;
+	let products;
+	let categories;
 
-	it('should pass readme example', () => {
+	beforeEach(() => {
 		// This is an array of data from server
-		let facts = [
+		facts = [
 			{ id: 1, region: 'North', year: 2017, month: 'January', product: 'Product 1', category: 'Category 1', value: 737 },
 			{ id: 2, region: 'South', year: 2017, month: 'April', product: 'Product 2', category: 'Category 1', value: 155 },
 			{ id: 3, region: 'West', year: 2018, month: 'April', product: 'Product 3', category: 'Category 2', value: 112 },
@@ -14,7 +21,7 @@ describe('readme', () => {
 		];
 
 		// This is the data dimensionHierarchies we need to obtain
-		let dimensionHierarchies = [
+		dimensionHierarchies = [
 			{
 				dimensionTable: {
 					dimension: 'regions',
@@ -44,25 +51,25 @@ describe('readme', () => {
 		];
 
 		// We send it all to the constructor
-		let cube = Cube.create(facts, dimensionHierarchies);
+		cube = Cube.create(facts, dimensionHierarchies);
 
-		const regions = [
+		regions = [
 			{ id: 1, region: 'North' },
 			{ id: 2, region: 'South' },
 			{ id: 3, region: 'West' }
 		];
-		const date = [
+		date = [
 			{ id: 1, year: 2017, month: 'January' },
 			{ id: 2, year: 2017, month: 'April' },
 			{ id: 3, year: 2018, month: 'April' }
 		];
-		const products = [
+		products = [
 			{ id: 1, product: 'Product 1', categories_id: 1 },
 			{ id: 2, product: 'Product 2', categories_id: 1 },
 			{ id: 3, product: 'Product 3', categories_id: 2 },
 			{ id: 4, product: 'Product 1', categories_id: 2 },
 		];
-		const categories = [
+		categories = [
 			{ id: 1, category: 'Category 1' },
 			{ id: 2, category: 'Category 2' },
 		];
@@ -73,12 +80,16 @@ describe('readme', () => {
 			{ id: 3, regions_id: 3, date_id: 3, products_id: 3, value: 112 },
 			{ id: 4, regions_id: 3, date_id: 3, products_id: 4, value: 319 },
 		];
+	});
 
+	it('check members', () => {
 		debug = isEqualObjects(regions, cube.getDimensionMembers('regions'));
 		debug = isEqualObjects(date, cube.getDimensionMembers('date'));
 		debug = isEqualObjects(products, cube.getDimensionMembers('products'));
 		debug = isEqualObjects(categories, cube.getDimensionMembers('categories'));
+	})
 
+	it('check getDimensionMembers', () => {
 		const result = cube.dice({ categories: [{ id: 1 }] });
 
 		debug = isEqualObjects(
@@ -86,7 +97,7 @@ describe('readme', () => {
 				{ id: 1, product: 'Product 1', categories_id: 1 },
 				{ id: 2, product: 'Product 2', categories_id: 1 },
 			],
-			cube.getDimensionMembersBySet('products', { categories: { id: 1 } })
+			cube.dice({ categories: { id: 1 } }).getDimensionMembers('products')
 		);
 
 		debug = isEqualObjects(
@@ -95,7 +106,7 @@ describe('readme', () => {
 				{ id: 3, product: 'Product 3', categories_id: 2 },
 				{ id: 4, product: 'Product 1', categories_id: 2 },
 			],
-			cube.getDimensionMembersBySet('products', { regions: [{ id: 2 }, { id: 3 }] })
+			cube.dice({ regions: [{ id: 2 }, { id: 3 }] }).getDimensionMembers('products')
 		);
 
 		debug = isEqualObjects(
@@ -103,22 +114,24 @@ describe('readme', () => {
 				{ id: 1, region: 'North' },
 				{ id: 2, region: 'South' },
 			],
-			cube.getDimensionMembersBySet('regions', { categories: { id: 1 } })
+			cube.dice({ categories: { id: 1 } }).getDimensionMembers('regions')
 		);
 
+	})
+	it('check getCells/getFacts', () => {
 		{
 			let set = { regions: { id: 1 }, date: { id: 1 }, products: { id: 1 } };
 			debug = isEqualObjects(
 				[
 					{ id: 1, value: 737, regions_id: 1, date_id: 1, products_id: 1 }
 				],
-				cube.getCellsBySet(set)
+				cube.dice(set).getCells()
 			);
 			debug = isEqualObjects(
 				[
 					{ id: 1, region: 'North', year: 2017, month: 'January', product: 'Product 1', category: 'Category 1', value: 737 }
 				],
-				cube.getFactsBySet(set)
+				cube.dice(set).getFacts()
 			);
 		}
 
@@ -129,14 +142,14 @@ describe('readme', () => {
 					{ id: 3, value: 112, regions_id: 3, date_id: 3, products_id: 3 },
 					{ id: 4, value: 319, regions_id: 3, date_id: 3, products_id: 4 },
 				],
-				cube.getCellsBySet(subSet)
+				cube.dice(subSet).getCells()
 			);
 			debug = isEqualObjects(
 				[
 					{ id: 3, region: 'West', year: 2018, month: 'April', product: 'Product 3', category: 'Category 2', value: 112 },
 					{ id: 4, region: 'West', year: 2018, month: 'April', product: 'Product 1', category: 'Category 2', value: 319 },
 				],
-				cube.getFactsBySet(subSet)
+				cube.dice(subSet).getFacts()
 			);
 		}
 
@@ -149,11 +162,11 @@ describe('readme', () => {
 					{ id: 3, value: 112, regions_id: 3, date_id: 3, products_id: 3 },
 					{ id: 4, value: 319, regions_id: 3, date_id: 3, products_id: 4 },
 				],
-				cube.getCellsBySet(emptySet)
+				cube.dice(emptySet).getCells()
 			);
 			debug = isEqualObjects(
 				facts,
-				cube.getFactsBySet(emptySet)
+				cube.dice(emptySet).getFacts()
 			);
 		}
 
@@ -164,17 +177,19 @@ describe('readme', () => {
 					{ id: 1, value: 737, regions_id: 1, date_id: 1, products_id: 1 },
 					{ id: 2, value: 155, regions_id: 2, date_id: 2, products_id: 2 },
 				],
-				cube.getCellsBySet(set)
+				cube.dice(set).getCells()
 			);
 			debug = isEqualObjects(
 				[
 					{ id: 1, region: 'North', year: 2017, month: 'January', product: 'Product 1', category: 'Category 1', value: 737 },
 					{ id: 2, region: 'South', year: 2017, month: 'April', product: 'Product 2', category: 'Category 1', value: 155 },
 				],
-				cube.getFactsBySet(set)
+				cube.dice(set).getFacts()
 			);
 		}
+	})
 
+	it('check addDimensionHierarchy', () => {
 		{
 			const facts = [
 				{ id: 1, product: 'TV', mark: 'Sony', country: 'China', count: 2 },